back/requirements.txt
Django==2.1.7
django-cors-headers==2.5.0
djangorestframework==3.9.2
